examples/science/af3-slurm/examples/simple_ipynb_launcher/requirements.txt
google-cloud-storage
google-cloud-secret-manager
jinja2
biopython
py3Dmol
plotly
pandas
